// src/core/hooks/useStrings/interfaces.ts
export enum Language {
  ptBR = 'ptBR',
  enUS = 'enUS',
}

export interface StringsType {
  components: {
    repositoryList: {
      notFounded: string;
    };
    searchBar: {
      button: string;
      placeHolder: string;
    };
    userCard: {
      subTitle1: string;
      subTitle2: string;
      button1: string;
      button2: string;
    };
    userNotFound: {
      warning: string;
    };
  };
}